+++ packages/public/frontend/docker/Dockerfile
@@ -3,9 +3,8 @@ COPY --from=serlo-org-workspace /usr/src/app/packages/public/frontend packages/p
 WORKDIR packages/public/frontend
 ENV NEXT_ASSET_PREFIX http://localhost:3000
 ENV ASSET_PREFIX http://localhost:3001
-ENV SERVE_ASSETS true
 
-ENTRYPOINT ["node", "serverless.js"]
+ENTRYPOINT ["node", "serverless/serverless.development.js"]
 EXPOSE 3000
 
 FROM development as release
