+++ packages/public/frontend/serverless/serverless.production.js
@@ -1,0 +1,7 @@
+const lib = require('./lib')
+
+lib
+  .createServer([lib.healthCheck, lib.renderPage, lib.page404])
+  .listen(3000, () =>
+    console.log('Frontend (production) on http://localhost:3000')
+  )
